Tuning hyper-params improvements 1
Added net info file. Better folder organisation of tuneHyper

+++ mod_nolearn/utils.py
@@ -22,7 +22,8 @@ def create_dir(dir):
 
 def copyDirectory(src, dest):
     try:
-        shutil.copytree(src, dest)
+        if not os.path.exists(dest):
+            shutil.copytree(src, dest)
     # Directories are the same
     except shutil.Error as e:
         print('Directory not copied. Error: %s' % e)
